+++ src/components/IcoBuilder.tsx
@@ -1,34 +1,4 @@
-import atomIco from "../static/img/atom.png";
-import nearIco from "../static/img/nearcoin.png";
-import tonIco from "../static/img/ton.png";
-import solIco from "../static/img/solana.png";
-import aurIco from "../static/img/aurora.png";
-import usnIco from "../static/img/usn.png";
-import ethIco from "../static/img/eth.png";
-
-const icons : any = {
-	sol: {
-		img: solIco
-	},
-	ton: {
-		img: tonIco
-	},
-	atom: {
-		img: atomIco
-	},
-	aurora: {
-		img: aurIco
-	},
-	eth: {
-		img: ethIco
-	},
-	near: {
-		img: nearIco
-	},
-	usn: {
-		img: usnIco
-	}
-}
+import icons from "../configs/configIco"
 
 export const icoBuilder = (chain: string) => {
 	let key: string = chain.toLocaleLowerCase();
