--- src/config.ts
@@ -1,3 +1,0 @@
-const arrOfChains: any = ['NEAR', 'SOL', 'USN', 'AURORA', 'ATOM', 'TON', 'ETH', 'wNEAR (TON)', 'wETH (TON)', 'wUSN (TON)', 'wATOM (TON)', 'wSOL (TON)', 'wAURORA (TON)']
-
-export default arrOfChains;
